backend/requirements.txt
# FastAPI and related packages
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9  # For handling file uploads
pydantic==2.6.1  # For data validation
email-validator==2.1.0  # For email validation

# Database
supabase==2.3.4  # Supabase Python client
sqlalchemy==2.0.27  # SQL toolkit and ORM
psycopg2-binary==2.9.9  # PostgreSQL adapter

# PDF Processing
PyMuPDF==1.23.26  # For PDF processing (fitz)
PyPDF2==3.0.1  # PDF processing utilities

# AI/ML
google-generativeai==0.3.2  # Gemini AI API
python-dotenv>=0.19.0

# Authentication and Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-dotenv==1.0.1  # Environment variables

# Background Tasks and Scheduling
celery==5.3.6  # For background tasks
redis==5.0.1  # Message broker for Celery
APScheduler==3.10.4  # For scheduling tasks

# Email and Notifications
emails==0.6  # For sending emails
jinja2==3.1.3  # For email templates

# Utilities
pandas==2.2.0  # For data manipulation
python-dateutil==2.8.2  # Date utilities
pytz==2024.1  # Timezone support
